+++ app-backend/db/src/test/scala/com/azavea/rf/database/ExportDaoSpec.scala
@@ -14,11 +14,15 @@ import org.scalatest._
 import io.circe._
 import io.circe.syntax._
 
-class ExportDaoSpec
-    extends FunSuite
-    with Matchers
-    with IOChecker
-    with DBTestConfig {
+class ExportDaoSpec extends FunSuite with Matchers with DBTestConfig {
 
-  test("types") { check(ExportDao.selectF.query[Export]) }
+  test("types") {
+    xa.use(
+        t => {
+          ExportDao.query.list.transact(t)
+        }
+      )
+      .unsafeRunSync
+      .length should be >= 0
+  }
 }