+++ app-backend/db/src/test/scala/com/azavea/rf/database/SceneWithRelatedDaoSpec.scala
@@ -64,7 +64,7 @@ class SceneWithRelatedDaoSpec
             } yield { (dbScenes1, listedScenes) }
 
             val (insertedScenes, listedScenes) =
-              scenesIO.transact(xa).unsafeRunSync
+              xa.use(t => scenesIO.transact(t)).unsafeRunSync
             val insertedNamesSet = insertedScenes.toSet map {
               (scene: Scene.WithRelated) =>
                 scene.name
@@ -112,7 +112,7 @@ class SceneWithRelatedDaoSpec
             } yield { (dbScenes, retrievedScenes) }
 
             val (insertedScenes, scenesInProject) =
-              scenesToIngestIO.transact(xa).unsafeRunSync
+              xa.use(t => scenesToIngestIO.transact(t)).unsafeRunSync
             val ingestableScenesIds = scenesInProject filter { scene =>
               (scene.statusFields.ingestStatus, scene.sceneType) match {
                 case (_, Some(SceneType.COG))       => false