+++ app-backend/db/src/test/scala/com/azavea/rf/database/UserGroupRoleDaoSpec.scala
@@ -29,7 +29,9 @@ class UserGroupRoleDaoSpec
     with PropTestHelpers {
 
   test("list user group roles") {
-    UserGroupRoleDao.query.list.transact(xa).unsafeRunSync.length should be >= 0
+    xa.use(t => UserGroupRoleDao.query.list.transact(t))
+      .unsafeRunSync
+      .length should be >= 0
   }
 
   test("insert a user group role") {
@@ -59,7 +61,7 @@ class UserGroupRoleDaoSpec
               )
             } yield { insertedUgr }
 
-            val insertedUgr = insertUgrIO.transact(xa).unsafeRunSync
+            val insertedUgr = xa.use(t => insertUgrIO.transact(t)).unsafeRunSync
 
             insertedUgr.userId == userCreate.id &&
             insertedUgr.groupType == ugrCreate.groupType &&
@@ -149,7 +151,8 @@ class UserGroupRoleDaoSpec
               )
             } yield { usersInPlatform }
 
-            val usersInPlatform = usersInPlatformIO.transact(xa).unsafeRunSync
+            val usersInPlatform =
+              xa.use(t => usersInPlatformIO.transact(t)).unsafeRunSync
 
             assert(
               usersInPlatform.count == 1,
@@ -238,7 +241,8 @@ class UserGroupRoleDaoSpec
               )
             } yield { usersInPlatform }
 
-            val usersInPlatform = usersInPlatformIO.transact(xa).unsafeRunSync
+            val usersInPlatform =
+              xa.use(t => usersInPlatformIO.transact(t)).unsafeRunSync
 
             assert(
               usersInPlatform.count == 2,
@@ -334,7 +338,8 @@ class UserGroupRoleDaoSpec
               )
             } yield { usersInPlatform }
 
-            val usersInPlatform = usersInPlatformIO.transact(xa).unsafeRunSync
+            val usersInPlatform =
+              xa.use(t => usersInPlatformIO.transact(t)).unsafeRunSync
 
             assert(
               usersInPlatform.count == 2,
@@ -447,7 +452,8 @@ class UserGroupRoleDaoSpec
               )
             } yield { usersInPlatform }
 
-            val usersInPlatform = usersInPlatformIO.transact(xa).unsafeRunSync
+            val usersInPlatform =
+              xa.use(t => usersInPlatformIO.transact(t)).unsafeRunSync
 
             assert(
               usersInPlatform.count == 2,
@@ -530,8 +536,9 @@ class UserGroupRoleDaoSpec
               }
             }
 
-            val (affectedRows, updatedUgr) =
-              updateUgrWithUpdatedAndAffectedRowsIO.transact(xa).unsafeRunSync
+            val (affectedRows, updatedUgr) = xa
+              .use(t => updateUgrWithUpdatedAndAffectedRowsIO.transact(t))
+              .unsafeRunSync
             // isActive should always be true since it doesn't exist on UserGroupRole.Creates and is set
             // to true when those are turned into UserGroupRoles
             affectedRows == 1 &&
@@ -585,8 +592,9 @@ class UserGroupRoleDaoSpec
               }
             }
 
-            val (affectedRows, updatedUgr) =
-              deactivateWithDeactivatedUgrIO.transact(xa).unsafeRunSync
+            val (affectedRows, updatedUgr) = xa
+              .use(t => deactivateWithDeactivatedUgrIO.transact(t))
+              .unsafeRunSync
             affectedRows == 1 &&
             updatedUgr.isActive == false &&
             updatedUgr.modifiedBy == userCreate.id &&
@@ -635,8 +643,9 @@ class UserGroupRoleDaoSpec
               }
             }
 
-            val updatedUGRs =
-              deactivateWithDeactivatedUgrIO.transact(xa).unsafeRunSync
+            val updatedUGRs = xa
+              .use(t => deactivateWithDeactivatedUgrIO.transact(t))
+              .unsafeRunSync
             updatedUGRs.size == 1 &&
             updatedUGRs
               .filter(